libs/pas-core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pas-core"
version = "0.1.0"
description = "Core logic for the PAS toolkit"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "keyring>=24.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pas_core"]
